feat: introduce CampusNavigatorViewModel + CampusNavigatorRepository for campus navigator screen
