+++ client-app/lib/features/campus_navigator/model/campus_navigation_model.dart
