+++ client-app/lib/features/campus_navigator/repository/campus_navigator_repository.dart
@@ -1,0 +1,9 @@
+import 'package:plan_sync/features/campus_navigator/model/campus_navigation_model.dart';
+
+abstract class CampusNavigatorRepository {
+  Future<List<CampusNavigationModel>> getLocations({
+    required int page,
+    required int limit,
+    String search,
+  });
+}
