+++ client-app/lib/features/campus_navigator/view/widgets/empty_campus_widget.dart
